Add files via upload
The process involved enabling surveys, creating and branding the survey, building survey pages, creating a new site, adding the survey to the site, configuring permissions for the guest profile, and making style adjustments. This approach demonstrates how powerful Salesforce's out-of-the-box capabilities can be in creating effective tools with minimal technical overhead. Explore the project to see how easy it is to deploy a fully functional employee survey using Salesforce configuration!

Here's the link to interact: https://shravyaportfolio-a-dev-ed.develop.my.site.com/survey/s/